Update checkstyle for switch statements
Traditional switch statements with case : have no indentation, and those with -> have 4 space indentation, according to the coding convention of the course.

The checkstyle is making noise for the switch statement with correct indentation in Task.java, because it has 4 space indentation instead of 0.

Add a suppression in suppressions.xml to fix this temporarily